CTF/2023/CTF-Cyber-jawara/Quals/Cryptography/matryoshka/requirements.txt
python-jose[cryptography]
fastapi
uvicorn